refactor: move  2fa-related logic into specific file
